some changes in page testimonial

+++ classtune/lib/delayed_batch_tranfer.rb
@@ -49,7 +49,7 @@ class DelayedBatchTranfer
     un_saved_batch = PdfSave.find_by_batch_id_and_status(@from,false)
     if !@students.blank? and (!saved_batch.blank? or un_saved_batch.blank?)
       
-      if @transfer_all == "Yes"
+      if @transfer_all == "Yes" && !saved_batch.blank?
         batch_saved = PdfSave.find(saved_batch.id)
         batch_saved.destroy
       end
